+++ OUDS/Core/OUDS/Tests/OUDSTheme/MockTheme/MockTheme+ColorSemanticTokens.swift
@@ -154,6 +154,10 @@ extension MockTheme {
 
     override var colorBackgroundBrandTertiary: ColorSemanticToken! { Self.mockThemeMultipleColorTokens }
 
+    // MARK: Semantic token - Colors - Background - Status - Neutral
+    
+    override var colorBackgroundStatusNeutral: ColorSemanticToken? { Self.mockThemeMultipleColorTokens }
+    
     // MARK: Semantic token - Colors - Background - Status - Attractive - Muted
 
     override var colorBackgroundStatusAttractiveMuted: ColorSemanticToken! { Self.mockThemeMultipleColorTokens }